Change our license from CC-BY-SA, to CC-BY
We originally licensed these icons under CC-BY-SA because we wanted to make them
available to you to incorporate into your apps, but we didn't want someone to be
able to package them up and resell them as an icon set.  We heard your concerns
about the scope of the share alike obligation.  Our primary goal here is to make
it easy for you to use the icons, so we've changed the license to CC-BY.