Change sprity/sprity-gm with spritesmith/gmsmith
When Installing the sprity library the build of the required library
lwip fails on newer environments.
Also the 'png-sprites' gulp task fails as consequence.

It seems lwip is an abandoned project at the moment.

That library can be replaced by the spritesmith toolkit which aims
to do the same task.

Also all relevant libraries are upgraded to the latest stable version.

links:
[lwip](https://github.com/EyalAr/lwip)
[spritesmith](https://github.com/Ensighten/spritesmith)